Got the publishing of the height map working as floats, terrain map data being sent to the footstep planner is not working

+++ ihmc-footstep-planning/src/visualizers/java/us/ihmc/footstepPlanning/ui/viewers/HeightMapVisualizer.java
@@ -94,7 +94,7 @@ public class HeightMapVisualizer extends AnimationTimer
    {
       /* Compute mesh */
       meshBuilder.clear();
-      IDLSequence.Integer heights = heightMapMessage.getHeights();
+      IDLSequence.Float heights = heightMapMessage.getHeights();
       double gridResolutionXY = heightMapMessage.getCellSizeInMeters();
       int centerIndex = HeightMapTools.computeCenterIndex(heightMapMessage.getWidthInMeters(), gridResolutionXY);
       double groundPlaneHeight = heights.min();
