+++ ihmc-perception/src/test/java/us/ihmc/perception/heightmap/HeightMapMessageToolsTest.java
@@ -27,12 +27,12 @@ public class HeightMapMessageToolsTest
 
       int centerIndex = HeightMapTools.computeCenterIndex(widthInMeters, cellResolution);
       int cellsPerAxis = 2 * centerIndex + 1;
-      Mat heightMap = new Mat(cellsPerAxis, cellsPerAxis, opencv_core.CV_16UC1);
+      Mat heightMap = new Mat(cellsPerAxis, cellsPerAxis, opencv_core.CV_32FC1);
       for (int i = 0; i < cellsPerAxis; i++)
       {
          for (int j = 0; j < cellsPerAxis; j++)
          {
-            heightMap.ptr(i, j).putShort((short) (i + j));
+            heightMap.ptr(i, j).putFloat((i + j));
          }
       }
 
@@ -45,7 +45,7 @@ public class HeightMapMessageToolsTest
          {
             for (int j = 0; j < cellsPerAxis; j++)
             {
-               assertEquals(heightMap.ptr(i, j).getShort(), heightMapResult.ptr(i, j).getShort());
+               assertEquals(heightMap.ptr(i, j).getFloat(), heightMapResult.ptr(i, j).getFloat());
             }
          }
          // Do nothing
